Some fixes to transform fragments + processor and update shading strategy to do simple deferred gbuffer pass

+++ engine/src/core/ecs/fragments/static_mesh_fragment.js
@@ -5,7 +5,7 @@ export class StaticMeshFragment extends Fragment {
 
     static initialize() {
         this.data = {
-            mesh: new Uint32Array(1),
+            mesh: new BigInt64Array(1),
             material_slots: new Uint32Array(this.material_slot_stride)
         };
     }
